+++ jdplus-main-base/jdplus-toolkit-base-parent/jdplus-toolkit-base-core/src/test/java/jdplus/toolkit/base/core/modelling/regular/tests/CanovaHansenForTradingDaysTest.java
@@ -17,12 +17,10 @@
 package jdplus.toolkit.base.core.modelling.regular.tests;
 
 import java.util.Arrays;
-import jdplus.toolkit.base.api.data.DoubleSeq;
 import jdplus.toolkit.base.api.dstats.RandomNumberGenerator;
 import jdplus.toolkit.base.api.timeseries.TsData;
 import jdplus.toolkit.base.api.timeseries.TsPeriod;
 import jdplus.toolkit.base.core.dstats.Normal;
-import jdplus.toolkit.base.core.modelling.regular.tests.CanovaHansenForTradingDays;
 import jdplus.toolkit.base.core.random.XorshiftRNG;
 import tck.demetra.data.Data;
 import org.junit.jupiter.api.Test;
@@ -53,10 +51,10 @@ public class CanovaHansenForTradingDaysTest {
 
     public static void main(String[] args) {
         Normal N = new Normal();
-        int M = 100000;
+        int M = 200000;
         int P = 12;
         RandomNumberGenerator rng = XorshiftRNG.fromSystemNanoTime();
-        for (int i = P * 5; i <= P * 100; i += P * 5) {
+        for (int i = P * 5; i <= P * 50; i += P * 5) {
             double[] a = new double[M];
             double[][] s = new double[7][];
             for (int j = 0; j < 7; ++j) {
@@ -78,17 +76,17 @@ public class CanovaHansenForTradingDaysTest {
                 s[6][j] = ch.testDerived();
             }
             Arrays.sort(a);
-            for (int j = 0; j < s.length; ++j) {
-                Arrays.sort(s[j]);
-                System.out.print(s[j][(int) (M * .9)]);
-                System.out.print('\t');
-                System.out.print(s[j][(int) (M * .95)]);
-                System.out.print('\t');
-                System.out.print(s[j][(int) (M * .99)]);
-                System.out.print('\t');
-                System.out.print(s[j][(int) (M * .999)]);
-                System.out.print('\t');
-            }
+//            for (int j = 0; j < s.length; ++j) {
+//                Arrays.sort(s[j]);
+//                System.out.print(s[j][(int) (M * .9)]);
+//                System.out.print('\t');
+//                System.out.print(s[j][(int) (M * .95)]);
+//                System.out.print('\t');
+//                System.out.print(s[j][(int) (M * .99)]);
+//                System.out.print('\t');
+//                System.out.print(s[j][(int) (M * .999)]);
+//                System.out.print('\t');
+//            }
             System.out.print(a[(int) (M * .9)]);
             System.out.print('\t');
             System.out.print(a[(int) (M * .95)]);